+++ releases.md
@@ -1,6 +1,6 @@
 # Releases
 
-## Unreleased
+## v1.6.0
 
   - Add support for `update-ca-trust` on Linux sytems.
   - Better command output.